:lipstick: :art: UI(Style) Formate and enhance the code structure of the form groups
